refactor: move repeated code to one file (same as below)
